+++ crates/node/build.rs
@@ -1,8 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-// Copyright (C) 2026 Robson Trasel
-
-extern crate napi_build;
-
-fn main() {
-    napi_build::setup();
-}